+++ bookops_worldcat/utils.py
@@ -3,7 +3,7 @@
 """
 Shared utilities module.
 """
-
+import re
 from typing import List, Union
 
 from .errors import InvalidOclcNumber
@@ -26,13 +26,18 @@ def prep_oclc_number_str(oclcNumber: str) -> str:
         `oclcNumber` as str
 
     Raises:
-        InvalidOclcNumber: If `oclcNumber` argument is not a str or int.
+        InvalidOclcNumber: If `oclcNumber` argument does not match
+        OCLC number formatting rules.
     """
 
-    if oclcNumber.strip().startswith("ocm") or oclcNumber.strip().startswith("ocn"):
+    if re.match(r"^ocm[0-9]{,8}$", oclcNumber.strip()) or re.match(
+        r"^ocn[0-9]{9}$", oclcNumber.strip()
+    ):
         oclcNumber = oclcNumber.strip()[3:]
-    elif oclcNumber.strip().startswith("on"):
+    elif re.match(r"^on[0-9]{10,}$", oclcNumber.strip()):
         oclcNumber = oclcNumber.strip()[2:]
+    elif re.match(r"^\(OCoLC\)[0-9]{8,}$", oclcNumber.strip()):
+        oclcNumber = oclcNumber.strip()[7:]
 
     try:
         oclcNumber = str(int(oclcNumber))
@@ -47,7 +52,7 @@ def verify_oclc_number(oclcNumber: Union[int, str]) -> str:
 
     Args:
         oclcNumber:
-            OCLC record number as string
+            OCLC record number as string or integer
 
     Returns:
         `oclcNumber` as str
@@ -92,7 +97,7 @@ def verify_oclc_numbers(
     elif isinstance(oclcNumbers, int):
         oclcNumbers_lst = _str2list(str(oclcNumbers))
     elif isinstance(oclcNumbers, list):
-        oclcNumbers_lst = oclcNumbers  # type: ignore
+        oclcNumbers_lst = [str(n) for n in oclcNumbers]
     else:
         raise InvalidOclcNumber(
             "Argument 'oclcNumbers' must be a single integer, a list or a "